Additional constraints for adding host when no mhost are present and for deletion of mhost when no other mhosts left

+++ monsys/src/main/java/edu/max/monsys/controller/MonitoringHostsController.java
@@ -47,8 +47,11 @@ public class MonitoringHostsController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> delete(@PathVariable("id") Integer id) {
-        monitoringHostRepository.deleteById(id);
-        return ResponseEntity.noContent().build();
+        if (monitoringHostRepository.count() == 1)
+            return new ResponseEntity<>("Сперва добавьте другой хост мониторинга", HttpStatus.BAD_REQUEST);
+        else
+            monitoringHostRepository.deleteById(id);
+        return ResponseEntity.ok("");
     }
